arch: Profile generation logic
We want to generate CPU profiles by running a CLI tool on the host
that is to be the profile's compatibility target. This commit prepares
for such a tool by introducing most of the generation logic, but the
necessary dependencies will be introduced in a follow up commit.

Signed-Off-by: Oliver Anderson <oliver.anderson@cyberus-technology.de>
On-behalf-of: SAP oliver.anderson@sap.com