arch: CPUID definitions: intel & kvm
We prepare for CPU profile generation by introducing the missing
dependencies: The CPUID definitions declaring what to do with the
CPUIDs we obtain from the compatibility target (the host CPU we want to
become live-migration compatible with).

Note that the "short", description and migration compatibility
requirement fields will not be used in this PR, but they are there to
help reviewers and auditors understand why certain profile policies
are set. Furthermore these additional fields could become useful in
the future. The migration compatibility requirements could for instance
be used to generate more thorough live migration checks, and the
short and long descriptions could be utilized for better diagnostics.

Signed-Off-by: Oliver Anderson <oliver.anderson@cyberus-technology.de>
On-behalf-of: SAP oliver.anderson@sap.com